requirements.txt
groq
streamlit
pytest
python-dotenv